redirect to error page done

+++ src/Controller/Back/MessageController.php
@@ -28,8 +28,11 @@ class MessageController extends AbstractController
     /**
      * @Route("/{id}", name="back_message_show", methods={"GET"})
      */
-    public function show(Message $message): Response
+    public function show(Message $message = null): Response
     {
+        if( $message === null) {
+            return $this->redirectToRoute('back_error');
+        }
         $message->setIsRead(true);
         $entityManager = $this->getDoctrine()->getManager();$entityManager = $this->getDoctrine()->getManager();
         $entityManager->flush();
@@ -42,8 +45,11 @@ class MessageController extends AbstractController
     /**
      * @Route("/{id}", name="back_message_delete", methods={"POST"})
      */
-    public function delete(Request $request, Message $message): Response
+    public function delete(Request $request, Message $message = null): Response
     {
+        if( $message === null) {
+            return $this->redirectToRoute('back_error');
+        }
         if ($this->isCsrfTokenValid('delete'.$message->getId(), $request->request->get('_token'))) {
             $entityManager = $this->getDoctrine()->getManager();
             $entityManager->remove($message);
